step4_storage_processing/requirements.txt
pymongo>=3.12.0
kafka-python>=2.0.2